Change the default index colors
Why changed?

Because the colors used were too simple for a game, it was not visually appealing and the variables used did not correspond correctly to their intended use.

What changed?

The color variables used were changed to ones containing gradient colors and a pair of dark and light colors, then these variables were used for the respective backgrounds and text, and the backgrounds were added the size cover property. These changes were made to the styles of the index